Code Push

+++ src/main/java/com/example/unimanagement/model/Admin.java
@@ -4,7 +4,7 @@ import jakarta.persistence.Entity;
 
 @Entity
 public class Admin extends User{
-    public Admin(long userId, String fName, String lName, String email, String password, UserRole role) {
+    public Admin(int userId, String fName, String lName, String email, String password, UserRole role) {
         super(userId, fName, lName, email, password, role);
     }
